backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
aiosmtplib
jinja2
